basic movements

+++ lib/position.rs
@@ -1,5 +1,5 @@
 #[derive(Serialize)]
 pub struct Position {
-  pub row: i32,
-  pub col: i32,
+  pub row: usize,
+  pub col: usize,
 }
